ci: Add OBS helper workflow
This workflow will simply add another check in a PR with the URL to the
OBS project that should be created by the OBS scm integration, so you
can click on it and delete it if it is broken and needs to get recreated.

Also it's quicker to check for already running tests, because OBS only
reports URLS when the build has finished.

I also increased the number of passing tests we require in mergify.
We never adapted the number when we added new tests.

Issue: https://progress.opensuse.org/issues/165144