video_stream: add support for ustreamer >= 6.10
Ustreamer 6.10 changed the frame format to v7. Formats v5 and v6 also
exist (they are very similar to v4), but I don't have the need to
support them.

The main difference is changed the header structure layout to be better
aligned. Now the header takes 128 bytes instead of 129. To support
different layouts, unpack the common header part first, and then save
offsets to remaining fields based on the header version.

The other difference recent ustreamer is using RGB3 encoding for the "no
signal" static image on raw sink, instead of JPEG encoding. Support for
this format was added in the previous commit.

Add relevant tests too. There are two tests for RGB3 encoding - one is
with the actual "no signal" message (something that ustreamer really
produces). But since it's grey, it does't really checks for byte order,
so add a second one with colorful image.

For now keep support for old ustreamer too.