Improve test for websocket error handling
When another webserver is running in the background, apache2 in my case, it might actually accept the connection instead of throwing an error as expected by the, now modfied, subtest.
This commit avoids this by providing a invalid IP and a unrelated unused port number (4), instead of just port number 0, which may be active at the time of test execution.