Revert incorrect fixes for qemu test coverage
This reverts commit b8f579fb534e45a193f910b27191dd8dc571b988
("Remove unreachable states in tests 18-backend-qemu") and
7c71d4ecbae0d234bf942e62193ae8d86233cf00 ("Fix code coverage
for qemu.pm"). The latter just marks stuff as unreachable,
which is the worst way to 'achieve' coverage, and it is possible
to do it better here (that will follow). The former is simply
incorrect; it incompletely undoes the changes from
86cd4419690e20ade4adcd84df5cd9d24e442036 under the incorrect
belief that they are unnecessary because we "know" the value of
the variable in question, but we do not. Its value depends on
the version of qemu installed on the system running the tests.
If you run the tests after that commit on a system with qemu
8 installed, they fail.

Subsequent commits will solve things correctly, by manipulating
the apparent qemu version so both codepaths are tested no matter
which qemu is installed on the system under test.