autotest/loadtest: Update test fullname along with name when duplicated
Currently, if the developer choose to take control of the test and pause at
a duplicate, the job will not pause. The issue is that
_handle_command_set_current_test in OpenQA/Isotovideo/CommandHandler.pm
needs the correct $full_test_name, which might not be accurate anymore. Fix
this by also updating the test "fullname" in loadtest. While at it, remove
"fullname" from the t/08-autotest.t test regarding duplicates.

Signed-off-by: Ricardo B. Marliere <rbm@suse.com>