Avoid endless loop in case of SSH read errors in `check_ssh_serial`
The `read2` call might time out (the default timeout is 5 minutes) or
otherwise fail. So far this leads to an endless loop until `eof` is reached
which might not happen in a timely manner, e.g. because the command being
executed is stuck. Then the whole backend would get stuck and the test
might run into `MAX_JOB_TIME`, see
https://progress.opensuse.org/issues/176076.

With this change the timeout on SSH operations is effective in case the
invoked command does not produce any output as tests would now die with
`Time out waiting for data (-9 LIBSSH2_ERROR_TIMEOUT)` when a command does
not return any new output for the timeout.

Since the default timeout is only 5 minutes this might disrupt long
operations such as copying big files over network if those operations don't
produce any output while running. So maybe it should be possible to set the
timeout on command-level (not included by this change).

This change also provokes more immediate errors in other cases where `read2`
fails.