Fix race condition when autotest terminates
When autotest terminates because all tests are done it sends the message
`tests_done` to isotovideo to indicate a normal exit. If isotovideo does
not get this message it concludes autotest has exited prematurely and
considers the test incomplete.

This makes generally sense. However, autotest exists immediately after
sending the `tests_done` message. Then isotovideo handles that exit by
setting its loop condition to false (like the it does when any of its
persistent child processes exists). This means it is not guaranteed that
the I/O loop of isotovideo still reads the `tests_done`. In case it does
not the test would be wrongfully considered incomplete.

This is happening from time to time in production openQA-in-openQA tests,
see https://progress.opensuse.org/issues/174259#note-24.

This change should fix the issue by letting autotest wait until isotovideo
responds to the `tests_done` message. It is actually normal for autotest
to wait for isotovideo to respond as this happens on every query_isotovideo
invocation. So it shouldn't cause any problems to use the same approach
also to communicate termination.