Use distinct and lower timeout for rsync call in svirt backend
The timeout for the rsync call is only exceeded when there is no I/O
activity at all for that time. So it is possible to use a lower timeout
here as suggested in review comments. This change therefore sets the rsync
timeout to 2.5 minutes and makes that configurable via
`SVIRT_ASSET_DOWNLOAD_INACTIVITY_TIMEOUT_M`. The timeout of 15 minutes for
the whole download (which was set in the previous commit) remains
unchanged.

Related ticket: https://progress.opensuse.org/issues/176868