Fix log colors
The problem was that when calling `fctinfo` for example, but the log level
would not include info, it would print the `color('yellow')`, but not
the `color('reset')`.

The result is that the prefix

    2025-02-13T15:50:34.660152Z] [debug] [pid:2866]

itself is not colored anymore, so I had to adapt the tests.

Issue: https://progress.opensuse.org/issues/177024