testapi: allow keeping mouse in place after clicking
Sometimes it's useful to keep the mouse cursor in place after clicking,
to avoid spurious hover events between actions. This is especially
useful when emulated mouse is really a mouse (with relative movement)
instead of a tablet device (with instant teleport).
On the API side, there is already mousehide arg, accepting 0 and 1
values (or true/false). Logically, it should be mousehide=>0, but that
has other meaning already. Implement this as mousehide=>-1 instead.