Revert removal of '%cmd' after regressions
This partially reverts "Remove unused variables and exports from
testapi" as %cmd is actually used in test distributions.

I consider the alternative and necessary change to osado for `%cmd` too
big and likely other test distributions are affected the same. But
`%vars` might be ok so updated the commit accordingly.

Also bumping testapi version to 44.