tinycv: scale starting search position if size is different
This is heuristic for reusing needles made on a different SUT
resolution. This assumes that objects that were at the center of the
screen will remain at the screen center if the resolution changes (and
similar for other screen areas). This works only if the object itself
doesn't change (for example its size is the same, etc).

The idea is to reuse needles made on XGA, on tests running at 4K. Some
sure needs to be re-made, but based on initial tests, many at least
partially match if the search area is adjusted. This didn't allowed
reusing the same needles so far, but made creating new needles much
easier. Unless HiDPI scaling kicks in, but that requires new needles
anyway.

If the scene and needle resolutions are the same, this is no-op
(multiplies by 1).