Try to increase pipe size for external video encoder
With 1024x768 frame, a single frame takes 2.3MB, which is bigger than
the default pipe size (64KB). The need to write in a couple iterations
(with waiting until the encoder retrieves the data in between) has little
performance impact in that case, and queing those frames takes limited
amout of memory. But with 3840x2160 frame, the difference is a lot
bigger (24MB > 64KB) and the impact is also significantly bigger. Slower
sending frames for encoding leads to a much higher memory consumption
and in some cases to the worker beeing killed by OOM-killer.

The default value for /proc/sys/fs/pipe-max-size is 1MB which is not
enough to increase even for 1024x768 frames. With 3840x2160 frames,
/proc/sys/fs/pipe-user-pages-soft limit can be reached too (64MB by
default, but applied to all pipes of a user combined). So, without
increasing global limits, increasing pipe size will fail.
Try to increase the pipe size, but when it fails, just emit a warning
(with suggested solution) instead of crashing.