Fix error from stricter Git permission checks when checking repo URL
This is a follow-up of bafb9efe512c50022d536abb42a86d578f369ce9 which
applies the same approach to the `git_remote_url` function as well.

This problem shows up in openQA-in-openQA tests where we see log messages
like these:

```
[2025-05-11T20:05:00.450213-04:00] [debug] [pid:12631] git hash in 'opensuse': 26bd263220ee8c5c78dee6649dca3a00c1f633b5
fatal: detected dubious ownership in repository at '/var/lib/openqa/share/tests/opensuse'
To add an exception for this directory, call:
	git config --global --add safe.directory /var/lib/openqa/share/tests/opensuse
[2025-05-11T20:05:00.458613-04:00] [debug] [pid:12631] git url in 'opensuse': "UNKNOWN (origin remote not found)"
```

So determining the Git hash works due to
bafb9efe512c50022d536abb42a86d578f369ce9 but determining the Git URL still
runs into the same problem.

I reproduced the problem and verified this fix locally by simply changing
the ownership of my tests directory so that it no longer matches the user I
run os-autoinst with.

Note that openQA-in-openQA tests are not the problem here. They execute
`fetchneedles` as root but this script handles that and the checkout is
nevertheless correctly owned by `geekotest`. The mismatch is only because
isotovideo runs as `_openqa-worker` but this is expected.

Related ticket: https://progress.opensuse.org/issues/183086