setup-multi-machine: extend network service detection
network.service being a symlink to the configured networking
service is a SUSE-ism; on Fedora, for e.g., network.service was
for the old network-scripts and is now mostly no longer used.
Extend the detection so that if network.service-based detection
doesn't work, we make a guess based on enabled services under
common names.

Signed-off-by: Adam Williamson <awilliam@redhat.com>