Cleanup qcow images before perform a copy from local storage
A simple approach to perform automatic cleanup of kvm assets produced by the
s390x kvm backend (svirt) in os-autoinst is to remove the existing one if it
is already existed.

When the `add_disk` is invoked and the test settings provides HDD_* the
`add_disk` is called passing the HDD_* variable with the path which obtain from
the openQA location. if the image is not found the process dies. At least in
the zkvm_add_disk. I guess this can leave old images in the hypervisor, which
eventually should be cleaned up from cleanup-openqa-assets cronjob.

When `create` is passed in the `add_disk` the $file construct an .img file
(openQA-SUT-x.img usually). This is the case when there is no HDD_*

https://progress.opensuse.org/issues/181154
Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>