Add variable to indicate IPv6 availability to tests
As we cannot ensure IPv6 availability on all hosts connected to o3 and
other openQA instances, we can only signal it to the test. The test can
then check for this variable and notify the user about a needed change
(e.g. add a worker-class requirement to the test, skip the test with a
warning, etc).