t/consoles-s3270.t: fix with IPC::Run 20250809.0
A change in IPC::Run 20250809.0 assumes several things are defined
in `finish()`. This is a safe assumption in normal usage because
they're defined by `harness()`, which is called by `start()`. But
in our tests, we mock out Run's `start()`, so they're not defined,
and our test that calls `finish()` blows up with undefined ref
errors. To fix this, let's have our mocked `start()` define the
necessary things.

Signed-off-by: Adam Williamson <awilliam@redhat.com>