VNC: cap the delay on key down events
See: https://bugzilla.redhat.com/show_bug.cgi?id=2392684
It's an issue in Fedora tests where encryption passphrase entry
at boot was failing (on aarch64 only, for some reason) because
characters were being doubled. Jeremy Linton figured out it seems
to be caused by this variable delay between key down and key up
events that's part of the 'typing speed limit' implementation.

We set the `type_string` max_interval arg to 1 for typing the
passphrase. That means we try to do one keypress per 1.0333...
seconds, which results in a just-over-0.25 second sleep between
key down and key up events, which acts like holding each key
down for just over 0.25 seconds. That seems to cause the repeated
characters.

This tweaks the code so the key down sleep is capped at 0.008
seconds (chosen because it's approx what you get with the default
TYPING_LIMIT of 30, and no max_interval arg, so it's had a lot of
real-world use). The rest of the requested delay is done as the
key up sleep. It's a cap not a hardcoded value to allow for very
fast typing.

This also makes the delay arg of map_and_send_key (and related
functions) be the *total* delay requested, not half of it. This
made the implementation easier, but note the other three things
that call into this path already acted as if it behaved this way
- they pass values that are correct as the *total* delay. So
this change should 'fix' those as well. The arg is renamed to
make this clear.

Signed-off-by: Adam Williamson <awilliam@redhat.com>